refactor:refactor CreateExerciseScreen.kt
